fix bug ---minsum应该为double

+++ main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 
-int min_sum = 99999999;
+double min_sum = 99999999;
 vector<double> min_path;
 vector<double> mArray;
 vector<Point> point;
